+++ CHANGELOG.md
@@ -14,6 +14,7 @@ For a full diff see [`1.0.0...main`][1.0.0...main].
 - Renamed `Json::encoded()` to `Json::toString()` ([#155]), by [@localheinz]
 - Inlined `Decoder` into `SchemaValidator` ([#157]), by [@localheinz]
 - Disallowed injection of `Validator` into `SchemaValidator` ([#158]), by [@localheinz]
+- Removed `Schema` ([#161]), by [@localheinz]
 
 ## [`1.0.0`][1.0.0]
 
@@ -41,5 +42,6 @@ For a full diff see [`dcd4cfb...1.0.0`][dcd4cfb...1.0.0].
 [#155]: https://github.com/ergebnis/json-schema-validator/pull/155
 [#157]: https://github.com/ergebnis/json-schema-validator/pull/157
 [#158]: https://github.com/ergebnis/json-schema-validator/pull/158
+[#161]: https://github.com/ergebnis/json-schema-validator/pull/161
 
 [@localheinz]: https://github.com/localheinz